+++ tests/test_list.py
@@ -4,6 +4,7 @@
 
 import nf_core.list
 
+import json
 import mock
 import os
 import pytest
@@ -16,19 +17,29 @@ from datetime import datetime
 class TestLint(unittest.TestCase):
     """Class for list tests"""
 
-    @mock.patch("json.dumps")
     @mock.patch("subprocess.check_output")
-    @mock.patch("nf_core.list.LocalWorkflow")
-    def test_working_listcall(self, mock_loc_wf, mock_subprocess, mock_json):
+    def test_working_listcall(self, mock_subprocess):
         """ Test that listing pipelines works """
-        nf_core.list.list_workflows()
+        wf_table = nf_core.list.list_workflows()
+        assert "rnaseq" in wf_table
+        assert "exoseq" not in wf_table
 
-    @mock.patch("json.dumps")
     @mock.patch("subprocess.check_output")
-    @mock.patch("nf_core.list.LocalWorkflow")
-    def test_working_listcall_json(self, mock_loc_wf, mock_subprocess, mock_json):
+    def test_working_listcall_archived(self, mock_subprocess):
+        """ Test that listing pipelines works, showing archived pipelines """
+        wf_table = nf_core.list.list_workflows(show_archived=True)
+        assert "exoseq" in wf_table
+
+    @mock.patch("subprocess.check_output")
+    def test_working_listcall_json(self, mock_subprocess):
         """ Test that listing pipelines with JSON works """
-        nf_core.list.list_workflows([], as_json=True)
+        wf_json_str = nf_core.list.list_workflows(as_json=True)
+        wf_json = json.loads(wf_json_str)
+        for wf in wf_json["remote_workflows"]:
+            if wf["name"] == "ampliseq":
+                break
+        else:
+            raise AssertionError("Could not find ampliseq in JSON")
 
     def test_pretty_datetime(self):
         """ Test that the pretty datetime function works """
@@ -56,7 +67,7 @@ class TestLint(unittest.TestCase):
             "name": "myWF",
             "full_name": "my Workflow",
             "description": "...",
-            "archived": [],
+            "archived": False,
             "stargazers_count": 42,
             "watchers_count": 6,
             "forks_count": 7,
@@ -129,7 +140,7 @@ class TestLint(unittest.TestCase):
             "name": "myWF",
             "full_name": "my Workflow",
             "description": "rna",
-            "archived": [],
+            "archived": False,
             "stargazers_count": 42,
             "watchers_count": 6,
             "forks_count": 7,
@@ -144,7 +155,7 @@ class TestLint(unittest.TestCase):
             "name": "myWF",
             "full_name": "my Workflow",
             "description": "dna",
-            "archived": [],
+            "archived": False,
             "stargazers_count": 42,
             "watchers_count": 6,
             "forks_count": 7,
@@ -159,3 +170,39 @@ class TestLint(unittest.TestCase):
         workflows_obj.remote_workflows.append(rwf_ex2)
 
         assert len(workflows_obj.filtered_workflows()) == 1
+
+    def test_filter_archived_workflows(self):
+        """
+        Test that archived workflows are not shown by default
+        """
+        workflows_obj = nf_core.list.Workflows()
+        remote1 = {"name": "myWF", "full_name": "my Workflow", "archived": True, "releases": []}
+        rwf_ex1 = nf_core.list.RemoteWorkflow(remote1)
+        remote2 = {"name": "myWF", "full_name": "my Workflow", "archived": False, "releases": []}
+        rwf_ex2 = nf_core.list.RemoteWorkflow(remote2)
+
+        workflows_obj.remote_workflows.append(rwf_ex1)
+        workflows_obj.remote_workflows.append(rwf_ex2)
+
+        filtered_workflows = workflows_obj.filtered_workflows()
+        expected_workflows = [rwf_ex2]
+
+        assert filtered_workflows == expected_workflows
+
+    def test_show_archived_workflows(self):
+        """
+        Test that archived workflows can be shown optionally
+        """
+        workflows_obj = nf_core.list.Workflows(show_archived=True)
+        remote1 = {"name": "myWF", "full_name": "my Workflow", "archived": True, "releases": []}
+        rwf_ex1 = nf_core.list.RemoteWorkflow(remote1)
+        remote2 = {"name": "myWF", "full_name": "my Workflow", "archived": False, "releases": []}
+        rwf_ex2 = nf_core.list.RemoteWorkflow(remote2)
+
+        workflows_obj.remote_workflows.append(rwf_ex1)
+        workflows_obj.remote_workflows.append(rwf_ex2)
+
+        filtered_workflows = workflows_obj.filtered_workflows()
+        expected_workflows = [rwf_ex1, rwf_ex2]
+
+        assert filtered_workflows == expected_workflows